# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"userPostIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"locationName": {
						"type": "string"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"email",
					"userPostIds",
					"locationName",
					"longitude",
					"latitude",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_User_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"userPostIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"locationName": {
						"type": "string"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UserLocationUpdate": {
				"properties": {
					"userEmailStr": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"locationName": {
						"type": "string"
					}
				},
				"required": [
					"userEmailStr",
					"latitude",
					"longitude",
					"locationName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserPost": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"post_content": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"comments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"reactions": {
						"type": "number",
						"format": "double"
					},
					"locationName": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"user_id",
					"username",
					"post_content",
					"created_at",
					"comments",
					"reactions",
					"locationName",
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sound-circle",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUserByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Fetch information about a given user",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostNewUserProfile",
				"responses": {
					"200": {
						"description": "Message indicating successful creation",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new user and store them in the DB",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									},
									"location": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"location",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/newLocation": {
			"post": {
				"operationId": "PostUserLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User_"
								}
							}
						}
					}
				},
				"description": "Update a user's location",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLocationUpdate"
							}
						}
					}
				}
			}
		},
		"/posts": {
			"get": {
				"operationId": "GetPostsWithComments",
				"responses": {
					"200": {
						"description": "Array of posts with comments",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserPost"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Fetch all posts with their associated comments",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostNewUserPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"postDataStr": {
										"type": "string"
									},
									"usernameStr": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}